Prevent stretching of the canvas
Using CSS to change the width and height of the canvas causes the grid
to be stretched. In order to enlarge the canvas without distorting the
grid or causing pixelation, you must set the height and width attributes
on the canvas element itself. Unfortunately, popcode throws an error if
you use the width and height attributes in HTML, so we need to set it in
Javascript.

The default parameters for width and height have been chosen to match
the default width and height of a canvas element.

Source:
http://stackoverflow.com/questions/4938346/canvas-width-and-height-in-html5